bug fix in templates

+++ Templates/acc_in_torque_up_to_3_planets/nbody_gravity.c
@@ -60,7 +60,7 @@ void nbodyCalcAccelerations()
     }
     ////////////////////////////////////
     /* Disk accretion */
-    int g_dm_planet_list[3] = {g_dm_planet1, g_dm_planet2, g_dm_planet3};
+    float g_dm_planet_list[3] = {g_dm_planet1, g_dm_planet2, g_dm_planet3};
     for (int l = 2; l < NB_N; l++)
     {
         double add_mp = g_dm_planet_list[l-2]; // important to know if g_dm is a mass or  a mass rate
